Index

InventoryManagement.SupplyChainEnv

InventoryManagement.calculate_backlog

InventoryManagement.calculate_reorder

InventoryManagement.check_policy_inputs!

InventoryManagement.connect_nodes!

InventoryManagement.create_logging_dfs

InventoryManagement.create_order!

InventoryManagement.enforce_inventory_limits!

InventoryManagement.exit_order!

InventoryManagement.fulfill_from_production!

InventoryManagement.fulfill_from_stock!

InventoryManagement.get_capacity_and_supply

InventoryManagement.get_expected_consumption

InventoryManagement.get_inventory_state

InventoryManagement.initialize_echelons!

InventoryManagement.initialize_inventories!

InventoryManagement.inventory_components

InventoryManagement.is_terminated

InventoryManagement.isproduced

InventoryManagement.log_unfulfilled_demand!

InventoryManagement.make_shipment!

InventoryManagement.place_orders!

InventoryManagement.reorder_policy

InventoryManagement.reset!

InventoryManagement.show_action

InventoryManagement.simulate_markets!

InventoryManagement.simulate_policy!

InventoryManagement.update_echelons!

InventoryManagement.update_inventories!

InventoryManagement.update_shipments!

Environment

Missing docstring.

Missing docstring for InventoryManagement.seed!. Check Documenter's build log for details.

InventoryManagement.create_logging_dfs — Function

create_logging_dfs(net::MetaDiGraph, nodes::Vector, arcs::Vector, mats::Vector, echelons::Dict)

Create DataFrames to log timeseries results.

Inventory Policy

Missing docstring.

Missing docstring for InventoryManagement.is_valid. Check Documenter's build log for details.

InventoryManagement.check_policy_inputs! — Function

check_policy_inputs!(
    reorder_point::Dict, policy_param::Dict, review_period::Union{Int, AbstractRange, Vector, Dict}, 
    min_order_qty::Union{Real, Dict}, mats::Vector, request_nodes::Vector, supply_nodes::Vector
)

Validate inputs to inventory policy.

InventoryManagement.get_inventory_state — Function

get_inventory_state(
    n::Int, mat::Symbol, policy_variable::Union{Dict,Symbol},
    state1_grp::GroupedDataFrame, state2_grp::GroupedDataFrame
)

Check the inventory state at node n for material mat.

InventoryManagement.get_expected_consumption — Function

get_expected_consumption(env::SupplyChainEnv, n::Int, mat::Symbol, action::NamedArray)

Get expected raw material consumption at the producer node for downstream requests.

InventoryManagement.calculate_reorder — Function

calculate_reorder(
    n::Int, mat::Symbol, state::Float64, MOQ::Real, 
    policy_type::Union{Dict, Symbol}, policy_param::Dict
)

Calculate reorder quantity for material mat at node n.

Reorder Actions

InventoryManagement.initialize_inventories! — Function

initialize_inventories!(x::SupplyChainEnv)

Initialize on-hand and pipeline inventories with those from the previous period.

InventoryManagement.place_orders! — Function

place_orders!(x::SupplyChainEnv, act::NamedArray)

Place inventory replenishment requests throughout the network.

InventoryManagement.exit_order! — Function

exit_order!(x, arcs)

Abort order placement.

InventoryManagement.calculate_backlog — Function

calculate_backlog(x::SupplyChainEnv, sup::Int, req::Int, mat::Symbol)

Calculate quantity of material mat backlogged on the arc (sup,req) (includes open orders).

InventoryManagement.create_order! — Function

create_order!(x::SupplyChainEnv, sup::Int, req::Int, mat::Symbol, amount::Real)

Create and log order.

InventoryManagement.fulfill_from_stock! — Function

fulfill_from_stock!(
    x::SupplyChainEnv, src::Int, dst::Int, mat::Symbol, 
    lead::Float64, supply_grp::GroupedDataFrame, pipeline_grp::Union{Missing,GroupedDataFrame}
)

Fulfill request from on-hand inventory.

InventoryManagement.fulfill_from_production! — Function

fulfill_from_production!(
    x::SupplyChainEnv, src::Int, dst::Int, mat::Symbol, 
    lead::Float64, supply_grp::GroupedDataFrame, pipeline_grp::GroupedDataFrame, capacities::Dict
)

Fulfill request by scheduling material production.

InventoryManagement.isproduced — Function

isproduced(x::SupplyChainEnv, n::Int, mat::Symbol)

Check if material mat is produced in node n.

InventoryManagement.get_capacity_and_supply — Function

get_capacity_and_supply(
    bom::NamedArray, rmat_names::Vector, cmat_names::Vector, 
    capacities::Dict, supply_grp::GroupedDataFrame
)

Get available capacity and material supply at producer.

InventoryManagement.make_shipment! — Function

make_shipment!(x::SupplyChainEnv, src::Int, dst::Int, mat::Symbol, accepted_amount::Float64, lead::Float64, pipeline_grp::GroupedDataFrame)

Schedule shipment of material.

InventoryManagement.log_unfulfilled_demand! — Function

log_unfulfilled_demand!(x::SupplyChainEnv, sup::Int, req::Int, mat::Symbol, accepted::Float64, order_row::DataFrameRow)

Calculate unfulfilled demand and reallocate order to next supplier in the priority list.

InventoryManagement.update_shipments! — Function

update_shipments!(x::SupplyChainEnv)

Update inventories throughout the network for arrived shipments.

InventoryManagement.enforce_inventory_limits! — Function

enforce_inventory_limits!(x::SupplyChainEnv)

Discard any excess inventory (exceeding the inventory capacity at each node).

InventoryManagement.update_inventories! — Function

update_inventories!(x::SupplyChainEnv)

Update inventory position and inventory level for all materials and nodes.

InventoryManagement.inventory_components — Function

inventory_components(
    x::SupplyChainEnv, n::Int, mat::Symbol, 
    pipeline_grp::GroupedDataFrame, onhand_grp::GroupedDataFrame, 
    orders_grp::GroupedDataFrame
)

Extract components to determine inventory level and position.

InventoryManagement.initialize_echelons! — Function

initialize_echelons!(x::SupplyChainEnv)

Initialize echelon stocks at the current period to 0.

InventoryManagement.update_echelons! — Function

update_echelons!(x::SupplyChainEnv, n::Int, mat::Symbol, ipos::Float64, ech_grp::GroupedDataFrame)

Update echelon stocks for current time period.

InventoryManagement.simulate_markets! — Function

simulate_markets!(x::SupplyChainEnv)

Open markets, apply material demands, and update inventory positions.

Utilities

Missing docstring.

Missing docstring for InventoryManagement.show_action. Check Documenter's build log for details.

Missing docstring.

Missing docstring for InventoryManagement.check_inputs. Check Documenter's build log for details.