Index

InventoryManagement.SupplyChainEnv

InventoryManagement.calculate_profit!

InventoryManagement.check_inputs

InventoryManagement.enforce_inventory_limits!

InventoryManagement.is_terminated

InventoryManagement.place_requests!

InventoryManagement.reorder_policy

InventoryManagement.reset!

InventoryManagement.show_action

InventoryManagement.simulate_markets!

InventoryManagement.simulate_policy!

InventoryManagement.update_production!

InventoryManagement.update_shipments!

Environment

InventoryManagement.SupplyChainEnv — Type

Supply Chain Simulation Environment Constructor

Fields:

network::MetaDiGraph: Supply chain network directed graph.

markets::Vector: Vector of market nodes where demand occurs.

producers::Vector: Vector of producer nodes where material transformation occurs.

distributors::Vector: Vector of distribution centers (excludes market nodes).

materials::Vector: Vector with the names of all materials in the system.

bill_of_materials::Array: Square matrix with BOM (rows = inputs, cols = ouputs); indices follow materials list; row with positive value is a co-product, row with negative value is a input (reactant).

inv_on_hand::DataFrame: Timeseries with on hand inventories @ each node.

inv_level::DataFrame: Timeseries with inventory level @ each node (on-hand minus backlog, if backlogging is allowed)

inv_pipeline::DataFrame: Timeseries with pipeline inventories on each arc.

inv_position::DataFrame: Timeseries with inventory positions @ each node (inventory level + placed replenishments).

replenishments::DataFrame: Timeseries with replenishment orders placed on each arc.

shipments::DataFrame: Temp table with active shipments and time to arrival on each arc.

production::DataFrame: Temp table with active material production commited to an arc and time to ship.

demand::DataFrame: Timeseries with realization of demand at each market, and amounts sold, unfulfilled demand, and backlog.

profit::DataFrame: Timeseries with profit @ each node.

reward::Float64: Final reward in the system (used for RL)

period::Int: Current period in the simulation.

num_periods::Int: Number of periods in the simulation.

discount::Float64: Time discount factor (i.e. interest rate).

options::Dict: Simulation options

backlog::Bool: Indicator if backlogging is allowed.

reallocate::Bool: Indicator if unfulfilled requests should be reallocated to alternate suppliers.

evaluate_profit::Bool: Indicator if the profit should be evaluated at each node

capacitated_inventory::Bool: Indicator if inventory limits should be enforced

seed::Int: Random seed.

InventoryManagement.reset! — Function

reset!(env::SupplyChainEnv)

Reset a SupplyChainEnv (empty all logging dataframes and set simulation time to 0).

InventoryManagement.is_terminated — Function

is_terminated(env::SupplyChainEnv)

Check if a simulation has terminated (i.e., has reached the maximum number of periods).

Inventory Policy

InventoryManagement.reorder_policy — Function

reorder_policy(env::SupplyChainEnv, param1::Dict, param2::Dict,
                    kind::Symbol = :rQ, review_period::Int = 1)

Apply an inventory policy to specify the replinishment orders for each material throughout the SupplyChainEnv.

InventoryManagement.simulate_policy! — Function

simulate_policy!(env::SupplyChainEnv, args...)

Step through a simulation using a specified reorder policy. args are the arguments that are passed to the reorder_policy function.

Reorder Actions

InventoryManagement.place_requests! — Function

place_requests!(x::SupplyChainEnv, act::Array, arcs::Vector)

Place inventory replenishment requests throughout the network.

InventoryManagement.update_production! — Function

update_production!(x::SupplyChainEnv)

Update completed production at each producer node and send to the arc that it was commited to.

InventoryManagement.update_shipments! — Function

update_shipments!(x::SupplyChainEnv)

Update inventories throughout the network for arrived shipments.

InventoryManagement.enforce_inventory_limits! — Function

enforce_inventory_limits!(x::SupplyChainEnv)

Discard any excess inventory (exceeding the inventory capacity at each node).

Missing docstring.

Missing docstring for InventoryManagement.update_position!. Check Documenter's build log for details.

InventoryManagement.simulate_markets! — Function

simulate_markets!(x::SupplyChainEnv)

Open markets, apply material demands, and update inventory positions.

InventoryManagement.calculate_profit! — Function

calculate_profit!(x::SupplyChainEnv, arrivals::DataFrame)

Calculate profit at each node in the network (sales - production cost - purchase costs - transportation costs - holding costs).

Utilities

InventoryManagement.show_action — Function

show_action(action::Vector{T} where T <: Real, env::SupplyChainEnv)

Visualize a replenishment order vector as a DataFrame indicating how much of which material is being requested on each arc.

InventoryManagement.check_inputs — Function

check_inputs(network::MetaDiGraph, nodes::Base.OneTo, arcs::Vector,
                mrkts::Vector, plants::Vector, mats::Vector, bom::Array,
                num_periods::Int)

Check inputs when creating a SupplyChainEnv.